requirements.txt
faiss-gpu==1.7.2
numpy==1.23.4
pandas==1.5.0
scikit-learn==1.1.2
torch==1.12.1
transformers==4.23.1
notebook